% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-summary.R
\name{summary.DiscreteTestResults}
\alias{summary.DiscreteTestResults}
\title{Summarizing Discrete Test Results}
\usage{
\method{summary}{DiscreteTestResults}(object, ...)
}
\arguments{
\item{object}{an object of class \link{DiscreteTestResults}, usually
produced by a call to one of the packages test functions, e.g.
\code{\link[=binom.test.pv]{binom.test.pv()}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A \code{\link[=DiscreteTestResultsSummary]{summary.DiscreteTestResults}} R6 class
object.
}
\description{
\code{summary} method for class \code{"DiscreteTestResults"}.
}
\examples{
obj <- binom.test.pv(0:5, 5, 0.5)
summary(obj)

}
