% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.R
\name{wilcoxon_test_pv}
\alias{wilcoxon_test_pv}
\alias{wilcox_test_pv}
\title{Wilcoxon sign rank test}
\usage{
wilcox_test_pv(
  x,
  mu = 0,
  alternative = "two.sided",
  exact = TRUE,
  correct = TRUE,
  digits_rank = Inf,
  simple_output = FALSE
)
}
\arguments{
\item{x}{numerical vector forming the sample to be tested or list
of numerical vectors for multiple samples.}

\item{mu}{numerical vector of hypothesised location(s).}

\item{alternative}{character vector that indicates the alternative hypotheses; each value must be one of \code{"two.sided"} (the default), \code{"less"} or \code{"greater"}.}

\item{exact}{logical value that indicates whether p-values are to be calculated by exact computation (\code{exact = TRUE}; the default) or by a continuous approximation.}

\item{correct}{logical value that indicates if a continuity correction is to be applied (\code{correct = TRUE}; the default) or not. Ignored, if \code{exact = TRUE}.}

\item{digits_rank}{single number giving the significant digits used to
compute ranks for the test statistics.}

\item{simple_output, }{logical value that indicates whether an R6 class object, including the tests' parameters and support sets, i.e. all observable p-values under each null hypothesis, is to be returned (see below).}
}
\value{
If \code{simple.output = TRUE}, a vector of computed p-values is returned.
Otherwise, the output is a \code{\link{DiscreteTestResults}} R6 class object, which
also includes the p-value supports and testing parameters. These have to be
accessed by public methods, e.g. \verb{$get_pvalues()}.
}
\description{
\code{wilcoxon_test_pv()} performs an exact or approximate Wilcoxon sign rank test
about the location of a population on a single sample. In contrast to
\code{\link[stats:wilcox.test]{stats::wilcox.test()}}, only the one-sample test is performed, but it is
vectorised and only calculates \emph{p}-values. Furthermore, it is capable of
returning the discrete \emph{p}-value supports, i.e. all observable \emph{p}-values
under a null hypothesis. Multiple tests can be evaluated simultaneously.
}
\details{
The parameters \code{x}, \code{mu} and \code{alternative} are vectorised. If \code{x} is a list,
they are replicated automatically to have the same lengths. In case \code{x} is
not a list, it is added to one, which is then replicated to the appropriate
length. This allows multiple hypotheses to be tested simultaneously.

In the presence of ties or observations that are equal to \code{mu}, computation
of exact \emph{p}-values is not possible. Therefore, \code{exact} is ignored in these
cases and \emph{p}-values are calculated by a normal approximation.

If \code{digits_rank = Inf} (the default), \code{\link[base:rank]{rank()}} is used to
compute ranks for the tests statistics instead of
\code{\link[base:rank]{rank}}(\code{\link[base:Round]{signif(., digits_rank)}})
}
\examples{
# Constructing
set.seed(1)
r <- rnorm(1000)

# Computation of exact two-sided p-values and their supports
results_ex  <- wilcox_test_pv(r)
raw_pvalues <- results_ex$get_pvalues()
pCDFlist    <- results_ex$get_pvalue_supports()

# Computation of normal-approximated one-sided p-values ("less") and their supports
results_ap  <- wilcox_test_pv(r, alternative = "less", exact = FALSE)
raw_pvalues <- results_ap$get_pvalues()
pCDFlist    <- results_ap$get_pvalue_supports()

}
\references{
Hollander, M. & Wolfe, D. (1973). \emph{Nonparametric Statistical Methods}. Third
Edition. New York: Wiley. pp. 40-55. \doi{10.1002/9781119196037}
}
\seealso{
\code{\link[stats:wilcox.test]{stats::wilcox.test()}}
}
