% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom-props.R
\name{binom_prop_test_pv}
\alias{binom_prop_test_pv}
\title{Unconditional two-sample Binomial test}
\usage{
binom_prop_test_pv(
  x,
  n,
  p = NULL,
  alternative = "two.sided",
  pooled = TRUE,
  exact = TRUE,
  correct = TRUE,
  simple_output = FALSE
)
}
\description{
Unconditional two-sample Binomial test
}
