% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-results.R
\name{DiscreteTestResults}
\alias{DiscreteTestResults}
\title{Discrete Test Results Class}
\description{
This is the class used by \code{DiscreteTests} to output more detailed
results, if the \code{simple.output} parameter is set to \code{FALSE}. All
data members are private to prevent causing inconsistencies by deliberate or
inadvertent changes. However, the results can be read by public methods.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiscreteTestResults-new}{\code{DiscreteTestResults$new()}}
\item \href{#method-DiscreteTestResults-get_pvalues}{\code{DiscreteTestResults$get_pvalues()}}
\item \href{#method-DiscreteTestResults-get_support_values}{\code{DiscreteTestResults$get_support_values()}}
\item \href{#method-DiscreteTestResults-get_support_indices}{\code{DiscreteTestResults$get_support_indices()}}
\item \href{#method-DiscreteTestResults-get_parameters}{\code{DiscreteTestResults$get_parameters()}}
\item \href{#method-DiscreteTestResults-summary}{\code{DiscreteTestResults$summary()}}
\item \href{#method-DiscreteTestResults-print}{\code{DiscreteTestResults$print()}}
\item \href{#method-DiscreteTestResults-clone}{\code{DiscreteTestResults$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscreteTestResults-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiscreteTestResults-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{DiscreteTestResults} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscreteTestResults$new(
  test_name,
  parameters,
  alternative,
  p_values,
  support_values,
  support_indices,
  data_name
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{test_name}}{a single character string with the name of the
test(s).}

\item{\code{parameters}}{a named list containing vectors of the tests
parameters. The first element is always a vector
of the observed values, all other list items
contain the UNIQE parameter combinations of the
tests.}

\item{\code{alternative}}{a single character string with the testing
alternative, e.g. "two.sided".}

\item{\code{p_values}}{a numeric vector of the p-values calculated by
each discrete testing scenario.}

\item{\code{support_values}}{a list of UNIQUE numeric vectors containing all
p-values the respective discrete test setting can
produce.}

\item{\code{support_indices}}{a list of numeric vectors containing the test
indices that indicates to which individual
testing scenario each unique support belongs.}

\item{\code{data_name}}{a single character string with the name of the
variable that contains the observed data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscreteTestResults-get_pvalues"></a>}}
\if{latex}{\out{\hypertarget{method-DiscreteTestResults-get_pvalues}{}}}
\subsection{Method \code{get_pvalues()}}{
Returns the computed p-values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscreteTestResults$get_pvalues()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscreteTestResults-get_support_values"></a>}}
\if{latex}{\out{\hypertarget{method-DiscreteTestResults-get_support_values}{}}}
\subsection{Method \code{get_support_values()}}{
Returns the unique test supports.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscreteTestResults$get_support_values()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscreteTestResults-get_support_indices"></a>}}
\if{latex}{\out{\hypertarget{method-DiscreteTestResults-get_support_indices}{}}}
\subsection{Method \code{get_support_indices()}}{
Returns the indices that indicate to which testing scenario each
unique support belongs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscreteTestResults$get_support_indices()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscreteTestResults-get_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-DiscreteTestResults-get_parameters}{}}}
\subsection{Method \code{get_parameters()}}{
Return the list of the test parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscreteTestResults$get_parameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscreteTestResults-summary"></a>}}
\if{latex}{\out{\hypertarget{method-DiscreteTestResults-summary}{}}}
\subsection{Method \code{summary()}}{
Returns the summary table which is printed by \code{print()}. For better
readability, the packages functions pass non-syntactic parameter names.
As a result, the returned summary table may have non syntactic column
names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscreteTestResults$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscreteTestResults-print"></a>}}
\if{latex}{\out{\hypertarget{method-DiscreteTestResults-print}{}}}
\subsection{Method \code{print()}}{
Returns the computed p-values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscreteTestResults$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{further arguments passed to \code{print.data.frame}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiscreteTestResults-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiscreteTestResults-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiscreteTestResults$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
