% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteTests.R
\docType{package}
\name{DiscreteTests-package}
\alias{DiscreteTests}
\alias{DiscreteTests-package}
\title{DiscreteTests}
\description{
This package includes vectorised functions for computing only p-values of
several discrete tests. Exact and approximate computation methods are
available. For exact p-values, multiple two-sided p-value procedures are
included.
}
\author{
\strong{Maintainer}: Florian Junge \email{diso.fbmn@h-da.de}

Authors:
\itemize{
  \item Christina Kihn
}

}
