% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnemar.R
\name{mcnemar.test.pv}
\alias{mcnemar.test.pv}
\title{McNemar's Test for Count Data}
\usage{
mcnemar.test.pv(
  x,
  alternative = c("two.sided", "less", "greater"),
  ts.method = c("minlike", "blaker", "absdist", "central"),
  exact = TRUE,
  correct = TRUE,
  simple.output = FALSE
)
}
\arguments{
\item{x}{an integer vector with four elements, a 2-by-2 matrix
or an integer matrix (or data frame) with four columns
where each line represents a 2-by-2 table to be tested,
i.e. a testing scenario.}

\item{alternative}{single character string that indicates the alternative hypothesis and must be one of \code{"two.sided"} (the default), \code{"less"} or \code{"greater"}.}

\item{exact}{logical value that indicates whether p-values are to be calculated by exact computation (\code{exact = TRUE}; the default) or by a continuous approximation.}

\item{correct}{logical vaue that indicates if a continuity correction is to be applied (\code{correct = TRUE}; the default) or not. Ignored, if \code{exact = TRUE}.}

\item{simple.output}{logical value that indicates whether an R6 class object, including the support sets, i.e. all attainable p-values of each testing scenario, is to be returned (see below).}
}
\value{
If \code{simple.output = TRUE}, a vector of computed p-values is returned.
Otherwise, the output is a \code{\link{DiscreteTestResults}} R6 class object. Results are
instead being accessed by public methods, e.g. \verb{$get_pvalues()}.
}
\description{
Performs McNemar's chi-square test or an exact variation of it to assess the
symmetry of rows and columns in a 2-by-2 contingency table. Only fourfold
tables can be processed, but multiple tables can be passed at once. The exact
version of the test is essentially an exact binomial test.
}
\details{
It can be shown that McNemar's test is a special case of the binomial test.
Therefore, its computations are handled by \code{\link[=binom.test.pv]{binom.test.pv()}}.
In contrast to that function, \code{mcnemar.test.pv} does not allow
specifying exact two-sided p-value calculation procedures. The reason is that
McNemar's exat test always test for a probability of 0.5, in which case all
these exact two-sided methods yield exactly the same results.
}
\examples{
# Constructing
S1 <- c(4, 2, 2, 14, 6, 9, 4, 0, 1)
S2 <- c(0, 0, 1, 3, 2, 1, 2, 2, 2)
N1 <- rep(148, 9)
N2 <- rep(132, 9)
F1 <- N1 - S1
F2 <- N2 - S2
df <- data.frame(S1, F1, S2, F2)

# Construction of exact p-values and their supports
results.ex  <- mcnemar.test.pv(df, simple.output = FALSE)
raw.pvalues <- results.ex$get_pvalues()
pCDFlist    <- results.ex$get_support_values()

# Construction of chisquare p-values and their supports
#results.cs  <- mcnemar.test.pv(df, exact = FALSE, simple.output = FALSE)
#raw.pvalues <- results.cs$get_pvalues()
#pCDFlist    <- results.cs$get_support_values()

}
\references{
Agresti, A. (2002). \emph{Categorical data analysis}, 2nd ed. New York: John
Wiley & Sons. pp. 350–354.
}
\seealso{
\code{\link[stats:mcnemar.test]{stats::mcnemar.test()}}, \code{\link[=binom.test.pv]{binom.test.pv()}}
}
