% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-results.R
\name{summary.DiscreteTestResults}
\alias{summary.DiscreteTestResults}
\title{Summarizing Discrete Test Results}
\usage{
\method{summary}{DiscreteTestResults}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"DiscreteTestResults"}, usually
produced by a call to one of the packages test functions, e.g.
\code{\link[=binom.test.pv]{binom.test.pv()}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{summary} method for class \code{"DiscreteTestResults"}.
}
\details{
Simply returns the results of the \code{get_summary_class()} method of the
underlying \code{object}.
}
\examples{
obj <- binom.test.pv(0:5, 5, 0.5, simple.output = FALSE)
summary(obj)

}
